Revert "Remove unused view from test url"
This reverts commit 91d312100caab99a6ed70d2c8346d379ae874d4b.
