--- frameworks/django/optic_django_unittest/tests/unittest_test_app/views.py
@@ -1,13 +1,0 @@
-from django.http import HttpResponse, JsonResponse
-
-
-def view1(request):
-    return HttpResponse("view1")
-
-
-def view2(request):
-    return HttpResponse("view2")
-
-
-def json_view(request):
-    return JsonResponse({"foo": "bar"})
